added client code

+++ src/Server.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char* argv[])
 {
-    int port(3000);
+    unsigned short port(3000);
 
     ServerManager server(port);
